requirements.txt
PyGithub==2.1.1
openai==1.69.0
aiohttp==3.11.4
pydantic==2.10.6
python-dotenv==1.0.1
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0 